Added output file stuff
We need to print out the information through an output file, so I worked on that a bit. There are no errors, but the program is having trouble reading in files so I don't know if the output file stuff works, but I'm still working on it. I figured at least it would look like we tried.